+++ packages/html-tools/package.js
@@ -1,14 +1,14 @@
 Package.describe({
   name: 'html-tools',
   summary: "Standards-compliant HTML tools",
-  version: '1.1.1',
+  version: '1.1.3',
   git: 'https://github.com/meteor/blaze.git'
 });
 
 Package.onUse(function (api) {
-  api.use('ecmascript@0.14.4');
-  api.use('htmljs@1.1.0');
-  api.imply('htmljs@1.1.0');
+  api.use('ecmascript@0.15.1');
+  api.use('htmljs@1.1.1');
+  api.imply('htmljs@1.1.1');
 
   api.export('HTMLTools');
   api.mainModule('main.js');
@@ -16,10 +16,10 @@ Package.onUse(function (api) {
 
 Package.onTest(function (api) {
   api.use('ecmascript');
-  api.use('tinytest@1.0.11');
+  api.use('tinytest@1.1.0');
 
   api.use('html-tools');
-  api.use('htmljs@1.1.0');
+  api.use('htmljs@1.1.1');
   api.use('blaze-tools'); // for `toJS`
 
   api.addFiles([