+++ packages/spacebars-compiler/package.js
@@ -1,21 +1,20 @@
 Package.describe({
   name: 'spacebars-compiler',
   summary: "Compiler for Spacebars template language",
-  version: '1.2.1',
+  version: '1.3.1',
   git: 'https://github.com/meteor/blaze.git'
 });
 
 Npm.depends({
-  'uglify-js': '2.7.5'
+  'uglify-js': '3.16.1'
 });
 
 Package.onUse(function (api) {
-  api.use('ecmascript@0.14.4');
-  api.use('underscore@1.0.9');
+  api.use('ecmascript@0.15.1');
 
-  api.use('htmljs@1.1.0');
-  api.use('html-tools@1.1.0');
-  api.use('blaze-tools@1.1.0');
+  api.use('htmljs@1.1.1');
+  api.use('html-tools@1.1.3');
+  api.use('blaze-tools@1.1.3');
 
   api.export('SpacebarsCompiler');
 
@@ -24,7 +23,6 @@ Package.onUse(function (api) {
 
 Package.onTest(function (api) {
   api.use([
-    'underscore',
     'ecmascript',
     'tinytest',
     'spacebars-compiler',